[NtType] Remove GetTypesSize and roll back code into LoadTypes
Since we do a pretty complicate dance in order to fill out a buffer with
ObjectTypesInformation data, we might as well do it once and use the
result right away.

The CFG of LoadTypes is clearly not simple, and moreso since we use a
'using' statement to automatically liberate local heap buffers.