[NtTypes] Correctly compute buffer size allocated for ObjectTypesInformation (#18)
* [NtTypes] Correctly compute buffer size allocated for ObjectTypesInformation

`ntdll.NtQueryObject` is one in a long list of natives API with a pretty
funny way to use it. It takes a buffer and a buffer size, and on success
fills out the bufffer and write the written length in a return_length
parameter.

However if the buffer is not big enough, the return status is set to
STATUS_INFO_LENGTH_MISMATCH inciting the developer to embiggen the input
buffer and re-attempt the call. However return_length parameter is also
sometimes set but it should not be relied upon as long as status is not
equal to STATUS_SUCCESS.

* [NtType] Remove GetTypesSize and roll back code into LoadTypes

Since we do a pretty complicate dance in order to fill out a buffer with
ObjectTypesInformation data, we might as well do it once and use the
result right away.

The CFG of LoadTypes is clearly not simple, and moreso since we use a
'using' statement to automatically liberate local heap buffers.

* [NtType] Simplify LoadType's CFG by explicitly throwing exceptions